% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.index.full.R
\name{make.index.full}
\alias{make.index.full}
\title{Make index of full formula}
\usage{
make.index.full(var.vec)
}
\arguments{
\item{var.vec}{vector of variables to include in the index.to.test data.frame}
}
\description{
This function takes a vector of variables and returns an index.to.test data.frame containing the full var.formula. var.formula is the part of the formula with the independent variables, i.e. that comes after "~", e.g. "A + B".
}
\examples{
# Load library
library(vegan)
data(mite)
data(mite.env)
abu <- mite
meta <- mite.env
var.vec <- colnames(mite.env)
index.to.test <- make.index.full(var.vec)
}
\keyword{formula}
\keyword{full.}
\keyword{index.to.test}
\keyword{variables}
