% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.index.varpart.R
\name{make.index.varpart}
\alias{make.index.varpart}
\title{Make index for unique and shared explained variations in the variation partition}
\usage{
make.index.varpart(var, var.vec)
}
\arguments{
\item{var}{variable to be tested.}

\item{var.vec}{vector of variables which contain variable var.}
}
\description{
This function takes a variable (var) and the vector of variables (var.vec), which must contain the variable given (var). The function returns index.to.test data.frame containing the formulas for calculating unique and shared on a dbRDA model.
}
\examples{
# Load library
library(vegan)
data(mite)
data(mite.env)
abu <- mite
meta <- mite.env
var.vec <- colnames(mite.env)
index.to.test <- NULL
for (var in var.vec){
  index.to.test <- rbind(index.to.test,
                            make.index.varpart(var, var.vec))
}
}
\keyword{formula.}
\keyword{index.to.test}
\keyword{shared}
\keyword{unique}
\keyword{variables}
