% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.model.R
\name{select.minimal.model}
\alias{select.minimal.model}
\title{Select minimal model}
\usage{
select.minimal.model(dis, meta, direction)
}
\arguments{
\item{dis}{distance matrix produced by vegan.}

\item{meta}{data.frame containing variables to be selected by ordistep.}

\item{direction}{direction to be used in ordistep. Default = "both".}
}
\description{
This function takes a distance matrix and a matrix containing variables data (metadata) and perform ordistep variable selection. It requires vegan and stringr.
}
\examples{
# Load library
library(vegan)
data(mite)
data(mite.env)
abu <- mite
meta <- mite.env
abu.std <- decostand(abu, "hellinger")
dis <- vegdist(abu.std)
#Select minimal model
selected.model <- select.minimal.model(dis, meta)
}
\keyword{model}
\keyword{select}
