% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.varpart.R
\name{test.varpart.wrap}
\alias{test.varpart.wrap}
\title{Wrap for testing varpart}
\usage{
test.varpart.wrap(dis, meta, var.formula, element, perm.n)
}
\arguments{
\item{dis}{distance matrix produced by vegan.}

\item{meta}{data.frame containing variables to be selected.}

\item{var.formula}{Character containing the part of the formula with the independent variables, i.e. that comes after "~", e.g. "A + B".}

\item{element}{String. If element = "full", P-value and F statistic is not produced. Default = "not.full".}

\item{perm.n}{Integer. Number of permutation to be performed in ANOVA.}
}
\description{
This function wrap perform varpart "manually", calculating dbRDA, anova and Rsquare. It takes a distance matrix and a data.frame containing variables data (metadata) and perform "manual" varpart. It requires vegan v 2.4-5. P-value and adjusted F are not produced for full formulas, i.e. formula without the Condition() function. This is specified in the element paramenter, element = "full". Otherwise, it assumes that the formula contains Condition(), and therefore P-value and adjusted F are produced. A list containing the model and statistics is produced.
First, a dbRDA model is built. ANOVA type III (by = "margin") is performed. Statistics and results are collected in a list.
}
\examples{
# Load library
library(vegan)
data(mite)
data(mite.env)
abu <- mite
meta <- mite.env
abu.std <- decostand(abu, "hellinger")
dis <- vegdist(abu.std)
res <- test.varpart.wrap(dis = dis,
meta = meta,
var.formula = "Substrate + Condition(WatrCont + SubsDens)",
element = "unique",
perm.n = 999)
}
\keyword{anova}
\keyword{multivariate}
\keyword{partition}
\keyword{test}
\keyword{variation}
\keyword{varpart}
